put arrow svg into a single asset and reformat the directory for faq
