+++ src/pages/index.js
@@ -3,11 +3,11 @@ import React from 'react';
 import styled from 'styled-components';
 
 import '../../node_modules/animate.css/animate.min.css';
-import '../components/layout.css';
+import '../css/layout.css';
 
 import BigLogoWithGridAnimated from '../components/bigLogo';
 import WelcomeAnimated from '../components/welcome';
-import FAQAnimated from '../components/faq';
+import FAQAnimated from '../components/faq/faq';
 import ScheduleAnimated from '../components/schedule';
 import SponsorBoxAnimated from '../components/sponsorBox';
